requirements.txt
streamlit
openai>=1.0.0
chromadb-client
python-dotenv
tiktoken
pypdf
langchain
langgraph
langchain_openai
boardgamegeek2
lxml